Update CMakeLists.txt and Fix a lot of issues: - fix codebase issues that have to do with wrong type casting

+++ mclib/include/mclib/protocol/packets/Packet.h
@@ -44,7 +44,7 @@ protected:
 
 public:
     Packet() noexcept 
-        : m_Id(0xFF), 
+        : m_Id((s8)0xFF), 
           m_ProtocolState(protocol::State::Play), 
           m_Connection(nullptr), 
           m_ProtocolVersion(protocol::Version::Minecraft_1_11_2)
